+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Changed
+- Use floor instead of truncate for WGS84 to NDS coordinate conversion (C++ and Python)
+
 ## [v0.5.0] - 2025-12-03
 
 ### Project